refactor: split the context writer, and stop it dropping capture flags
write-context.py had grown to 1,860 lines doing about a dozen jobs. The size
already cost us three times: it hid the fold bug fixed in the previous commit,
our own adoption pass gave up and wrote about forty helpers up from their
comments rather than their code, and its single if/elif dispatch ladder caused
a silent data-loss bug of its own.

That bug is fixed here too. Passing --decision and --verified in one call
recorded the decision, printed "Recorded 1 decision(s)", dropped the
verification, and exited 0 — so nothing downstream could tell half the data had
vanished. Every capture flag in one call now takes effect and is reported.

Nothing a user runs changes. Every flag behaves identically, and no command
body, hook, preset, or calling script needed editing.

Dev details
-----------
Five new siblings, in a one-way dependency order:

  spec_context.py     the store — read/atomic write, feature-dir resolution,
                      history log and predicates, git helpers, vocabulary
  spec_deltas.py      the delta grammar and parser; pure, no file access
  capture.py          the additive capture writers
  task_sync.py        task markers, checkboxes, the per-task journal
  living_spec_fold.py apply_deltas, fold_living_spec, target resolution

The issue proposed four. spec_context.py is the fifth because the fold and the
capture writers both need the context read/write primitives — leaving those in
write-context.py forces an import cycle Python does not survive for a module
loaded by a hyphenated filename. write-context.py keeps the CLI, the step
lifecycle, journal finish/advance, terminal promotion, and the no-regress
guard: 618 lines, down from 1,926.

write-context.py re-exports every moved name. Four things outside it import it
as a module and call its internals — derive-from-files.py, status-context.py,
both Python test suites, and the living-spec eval check — so re-export is what
makes the no-caller-edits constraint achievable. All 74 names the pre-split
module exposed still resolve; verified by dir() comparison against a staged
pre-change copy.

The siblings are imported by plain name (with the script's own directory pushed
onto sys.path) rather than loaded dynamically, because package-manifest.py
derives the packing list by following plain imports. A dynamic load would be
invisible to that gate and the release archive would ship broken, which is the
exact class of failure #432 was.

Dispatch is now two groups. Capture flags (--set, --decision, --verified,
--concern, --expectation, --context, --coverage-req, --step-summary,
--living-specs, --fold-living-spec, --classification) are additive and all run,
each printing its own line; --classification is validated before anything is
written so a bad value records nothing and still exits 2. Lifecycle modes
(--tasks-file, --mark-complete, --finish, --advance, --materialize, --task, and
the default step update) keep first-match-wins — they are alternative readings
of one invocation, not composable writes. The default step update stays
suppressed whenever a capture flag is present.

test_cli_parity.py proves the interface held: 34 invocations covering every
dispatch path, run against a staged pre-change copy of the scripts directory and
compared on exit code, stdout, stderr, and the resulting context file. All 34
identical.
