+++ docs/architecture.md
@@ -15,6 +15,8 @@ The extension splits into three runtime layers and one configuration surface:
 
 The repo also ships a **separate, independently-versioned spec-kit extension** under `speckit-extension/` (its own README/CHANGELOG/`extension.yml`). Its Companion commands are assembled from composable **nodes** by `scripts/assemble-nodes.py` (with `scripts/companion_config.py` as the executable spec of the `.specify/companion.yml` hook/recipe contract); see [`speckit-extension/docs/node-model.md`](../speckit-extension/docs/node-model.md). The VS Code extension does not depend on these at runtime — it only dispatches command text the host resolves.
 
+Its capture scripts split by responsibility, in a one-way dependency order: `spec_context.py` (the `.spec-context.json` store, feature-dir resolution, the history log, the canonical vocabulary) and `spec_deltas.py` (the requirement-delta grammar) sit at the bottom with no local dependencies; `capture.py` (the additive capture writers) and `task_sync.py` (task markers, checkboxes, the per-task journal) build on the store; `living_spec_fold.py` (the fold-back) builds on all three; and `write-context.py` sits on top holding the command line, the step lifecycle, journal finish/advance, terminal promotion, and the no-regress guard. `write-context.py` re-exports every sibling name, so `derive-from-files.py`, `status-context.py`, the Python test suites, and the eval skills all keep importing one module. Any new sibling must be added to `package-manifest.py`'s `RUNTIME_SCRIPTS` and imported by plain name — the packaging gate derives what ships by following plain imports, so a dynamically-loaded module would ship missing. Full reference: [`capture-and-timing.md`](./capture-and-timing.md).
+
 ## Extension host (`src/`)
 
 The host code organises around four directories, each with a single responsibility.
