+++ speckit-extension/CHANGELOG.md
@@ -22,6 +22,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/); this ext
 - **The drift check no longer passes itself on a CI runner that lacks the history to check.** Most CI providers clone only the most recent slice of git history. On a one-commit checkout the drift check used to print a clean all-clear having compared against nothing; with a few commits it reported drift against whatever commit happened to be the oldest one available, so the list of changed files was wrong. Both outputs looked perfectly healthy. Drift now recognizes that its baseline is out of reach, skips those capabilities with `spec history unreachable (shallow clone)`, and tells you to fetch the full history (for example `actions/checkout` with `fetch-depth: 0`). A capability whose spec was committed inside the available slice is still checked as usual, a full clone is unaffected, and the exit code stays `0` — the check is advisory and never fails a build.
 - **A partly-skipped drift run no longer leads with a checkmark that speaks for capabilities it never looked at.** With two of nine capabilities checked, the report used to say `✓ All 2 checked capabilities in sync` with the seven skips as a footnote underneath. It now says `✓ 2 of 9 capabilities in sync; 7 not checked — <reason>` on one line. A run where everything was checked keeps its original wording.
 - **A repository whose history can't be read now says that, instead of blaming a missing file.** The drift check reported every capability as `spec.md not yet committed`, which sent you looking for a file that was there all along. It now reports `spec history unreadable`. A repository that simply has no commits yet still reports `spec.md not yet committed`, which is what is actually true of it.
+- **Folding the same spec changes twice no longer duplicates sections.** Folding a finished feature's requirement changes into a living spec is meant to be repeatable — run it again and nothing should move. It wasn't, whenever one fold both added a section and renamed or edited that same section. Add-and-rename was the damaging one: every fold appended another copy of the requirement, without limit, so five folds left five copies in a document the team treats as durable. Add-and-edit was milder but noisier — the section's body flipped between two versions on every fold, and it made the eval harness report a repeatability failure for input that was perfectly legitimate, which teaches everyone to ignore that check. Every combination of change verbs is now repeatable, and a re-fold is a genuine no-op.
+- **Recording a decision and a verification in one go no longer loses the verification.** Passing `--decision` and `--verified` together recorded the decision, reported "Recorded 1 decision(s)", dropped the verification, and exited successfully — so nothing downstream could tell that half the data had vanished. Every capture flag in a single call now takes effect, and each one is reported. `--set`, `--concern`, `--expectation`, `--context`, `--coverage-req`, `--step-summary`, `--living-specs`, `--fold-living-spec`, and `--classification` are all in this group. The lifecycle flags are unchanged.
 - **The drift check no longer reports "all in sync" after checking nothing.** Right after you adopt a batch of capabilities, none of their specs are committed yet, so every one is skipped — drift needs a committed baseline to diff against. The run used to end on a green all-clear anyway, which read as a clean bill of health for a check that never ran. It now ends on `0 checked, 9 skipped (spec.md not yet committed)`, and the `✓ All N checked capabilities in sync.` line is reserved for a run that genuinely examined at least one capability and found it clean. A partly-skipped run states both numbers. The `--json` output gained a `checked` count so a script can tell "clean" from "did not run", and the exit code stays `0` throughout — a skip is correct behavior, not a failure.
 - **Folding a spec back into a living spec now reports what it applied, not what it attempted.** A change block naming a requirement heading the living spec doesn't have is skipped, but the summary still counted it, so `~3 modified` printed when all three had quietly matched nothing. The counts now reflect what actually landed, and dropped changes are called out.
 - Living-spec discovery now stops at nested projects. If your repo contains sample apps, fixtures, or sandboxes that carry their own `.specify/companion.yml`, they are separate projects: their spec files no longer show up as orphans in the parent repo, and they are no longer invented into capabilities of it. A nested project that has living specs turned off is left alone entirely, so opting a sandbox out finally means nothing happens to it.
@@ -29,6 +31,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/); this ext
 - Two discovered specs in like-named folders no longer collapse onto one name, and a discovered spec never takes over the name of a capability you configured.
 
 ### Changed
+
+- **The script that records a spec's progress is now five focused files instead of one 1,800-line one.** Nothing you run changes — every flag behaves identically, and no command, hook, or setting needed editing. The size was the problem: it hid the fold bug above, it caused the dropped-flag bug above, and when the adoption wizard tried to describe it, it read only the surface and wrote about forty helpers up from their comments rather than their code. Delta parsing, the living-spec fold, task syncing, and capture each have their own file now, so the next bug is findable by name.
 - **Eight commands renamed so their names say what they belong to.** The living-specs commands now carry the family: `/speckit.companion.adopt` → `living-adopt`, `drift` → `living-drift`, `coverage` → `living-coverage`, `relocate` → `living-move`. The four hook commands are now named after the lifecycle event they handle rather than the generic "capture": `capture` → `after-specify`, `capture-plan` → `after-plan`, `capture-tasks` → `after-tasks`, `capture-implement` → `after-implement`. The hooks still auto-run, so nothing changes unless you invoked one of the four living-specs commands by hand. The other nine (`specify`, `plan`, `tasks`, `implement`, `auto`, `classify`, `mark-complete`, `status`, `resume`) keep their names — they mirror stock Spec Kit, and that parallel is the point.
 
   Entries below this one predate the rename and use the names as they were at the time. **Upgrading requires `specify extension remove companion` before reinstalling** — nothing prunes emitted commands, so a plain reinstall leaves the old names live alongside the new ones (see #461).
